Fix crash when using continue_path in train_gpt_xtts.py
Using continue_path in recipes/ljspeech/xtts_v2/train_gpt_xtts.py causes a crash because the XttsAudioConfig and XttsArgs classes are missing attributes that are saved in the config.json file but are discarded by the loader. This commit addresses the issue by ensuring these attributes are present to prevent crashes.